[FEAT] 마이 프로필  API

+++ src/main/java/com/follow_me/running_mate/domain/course/dto/request/CourseRequest.java
@@ -57,7 +57,7 @@ public class CourseRequest {
         private Double distance;
 
         @NotNull(message = "평균 페이스를 입력해주세요.")
-        private Integer averagePace;
+        private Double averagePace;
 
         @NotNull(message = "소모 칼로리를 입력해주세요.")
         private Integer caloriesBurned;
